requirements.txt
openai 
pydantic 
python-dotenv 
watchdog 
textual 
docker 
rich
fastapi
pymongo